Temporarily disabled git pull/checkout in install
There are two different scenarios for when app:install is
executed: a completely new project (no remote git repositories)
and a deployment of an existing project (everything exists).

Forcing git pull in the first scenario, now completely fails the
first install, as there is no remote repository yet.  To solve
this we'll need something like 'FORCE_GIT' options or another
solutions, which I'll work on in a bit.

For now, commenting out should do the trick.