Quote parameters in wp.cli.install script template (task #2646)
The primary reason for this commit is WP_TITLE variable, which is
often configured to include more than one word.  When passed
unquoted to the parameters in the bash script, spaces break things
making it appear like multiple parameters.

Quoting should solve this issue.

As a precaution, other variables were quoted as well.  Double
quotes were used which might come handy for extrapolating
environment/shell variables, but, at the same time, might cause
other issues (if, for example, there is a double quote in the
WP_TITLE).