Re-adding vlucas/phpdotenv v1.1+ (task #4292)
A lot of projects built on top of this project-template are
working with the .env and .env.example files using the library
that we utilized from the beginning - vlucas/phpdotenv.

Our Robo-based setup does not require this library anymore, due
to its own implementation, but in order to maintain some backward
compatibility, we'll leave it in for a while.