Disabled strict validation for composer (task #6236)
Using strict validation in composer is usually a good idea.  However,
we are pulling a lot of dependecies from the wpackagist.org, which
does not properly handle the resolution of the package versions.  In
order to work around it, we lock dependencies to very specific versions.
This, in turn, throws warnings with strict composer validation.